Add Umami and Mermaid plugins
