+++ contributing.md
@@ -4,7 +4,12 @@ Thank you for investing your time in contributing to my project!
 
 ## VuePress plugins
 
-This repository contains some plugins for VuePress v2.
+This repository contains few plugins for VuePress v2:
+
+- [vuepress-plugin-mermaid-js](https://github.com/azat-io/azat-io/tree/main/plugins/mermaid-js): Plugin for using Mermaid.js
+- [vuepress-plugin-open-graph](https://github.com/azat-io/azat-io/tree/main/plugins/open-graph): Plugin for generating open graph meta tags
+- [vuepress-plugin-remove-html-extension](https://github.com/azat-io/azat-io/tree/main/plugins/remove-html-extension): Plugin for generating clean urls
+- [vuepress-plugin-umami-analytics](https://github.com/azat-io/azat-io/tree/main/plugins/umami-analytics): Plugin for using Umami analytics
 
 If you spot a problem with one of my plugins, search if an issue already exists. If a related issue doesn't exist, you can open a [new issue](https://github.com/azat-io/azat-io/issues/new).
 
@@ -12,7 +17,7 @@ Also you can help me to solve [existing issue](https://github.com/azat-io/azat-i
 
 ## Article translations
 
-I will be very happy if you help me to translate my articles into any language.
+I will be very happy if you help me to translate articles of my [blog](https://azat.io) into any language.
 
 All articles are stored in the `content/` folder. This directory contains folders with articles named according to the [ISO-639-1](https://en.wikipedia.org/wiki/List_of_ISO_639-1_codes) code according to the languages of the articles.
