+++ package.json
@@ -18,7 +18,7 @@
     "@vuepress/plugin-shiki": "2.0.0-beta.51",
     "@vuepress/plugin-theme-data": "2.0.0-beta.51",
     "@vuepress/utils": "2.0.0-beta.51",
-    "autoprefixer": "10.4.9",
+    "autoprefixer": "10.4.10",
     "chokidar": "3.5.3",
     "iso-639-1": "2.1.15",
     "jsonc-parser": "3.2.0",
@@ -28,7 +28,6 @@
     "postcss": "8.4.16",
     "postcss-100vh-fix": "1.0.2",
     "reading-time": "1.5.0",
-    "rss": "1.2.2",
     "shiki": "0.10.1",
     "sitemap": "7.1.1",
     "tsm": "2.2.2",
@@ -37,8 +36,10 @@
     "vue": "3.2.39",
     "vue-router": "4.1.5",
     "vue3-click-away": "1.2.4",
+    "vuepress-plugin-mermaid-js": "workspace:1.0.0",
     "vuepress-plugin-open-graph": "workspace:1.0.0",
     "vuepress-plugin-remove-html-extension": "workspace:1.0.0",
+    "vuepress-plugin-umami-analytics": "workspace:1.0.0",
     "vuepress-vite": "2.0.0-beta.51"
   },
   "devDependencies": {
@@ -46,10 +47,10 @@
     "@logux/stylelint-config": "0.12.0",
     "@size-limit/file": "8.1.0",
     "@types/mermaid": "8.2.9",
-    "@types/node": "18.7.16",
+    "@types/node": "18.7.17",
     "@types/rss": "0.0.29",
-    "@typescript-eslint/eslint-plugin": "5.36.2",
-    "@typescript-eslint/parser": "5.36.2",
+    "@typescript-eslint/eslint-plugin": "5.37.0",
+    "@typescript-eslint/parser": "5.37.0",
     "eslint": "8.23.1",
     "eslint-config-standard": "17.0.0",
     "eslint-plugin-import": "2.26.0",
