+++ plugins/mermaid-js/client/index.ts
