+++ plugins/mermaid-js/readme.md
@@ -1,0 +1,34 @@
+# vuepress-plugin-mermaid-js
+
+<img src="https://avatars.githubusercontent.com/u/48539483?s=100" align="right" alt="VuePress" />
+
+![Version](https://img.shields.io/npm/v/vuepress-plugin-mermaid-js.svg)
+![Downloads](https://img.shields.io/npm/dm/vuepress-plugin-mermaid-js.svg)
+
+VuePress plugin for using Mermaid.
+
+## Installation
+
+```sh
+npm install --save vuepress-plugin-mermaid-js
+```
+
+## Usage
+
+Add plugin to your VuePress config:
+
+```js
+import { mermaidJsPlugin } from 'vuepress-plugin-mermaid-js'
+
+export default {
+  plugins: [mermaidJsPlugin()],
+}
+```
+
+## Contribution
+
+See [Contributing Guide](https://github.com/azat-io/azat-io/blob/main/contributing.md).
+
+## License
+
+MIT
