+++ plugins/open-graph/readme.md
@@ -1,5 +1,7 @@
 # vuepress-plugin-open-graph
 
+<img src="https://avatars.githubusercontent.com/u/48539483?s=100" align="right" alt="VuePress" />
+
 ![Version](https://img.shields.io/npm/v/vuepress-plugin-open-graph.svg)
 ![Downloads](https://img.shields.io/npm/dm/vuepress-plugin-open-graph.svg)
 
@@ -28,3 +30,11 @@ export default {
   ],
 }
 ```
+
+## Contribution
+
+See [Contributing Guide](https://github.com/azat-io/azat-io/blob/main/contributing.md).
+
+## License
+
+MIT
