+++ plugins/remove-html-extension/readme.md
@@ -1,5 +1,7 @@
 # vuepress-plugin-remove-html-extension
 
+<img src="https://avatars.githubusercontent.com/u/48539483?s=100" align="right" alt="VuePress" />
+
 ![Version](https://img.shields.io/npm/v/vuepress-plugin-remove-html-extension.svg)
 ![Downloads](https://img.shields.io/npm/dm/vuepress-plugin-remove-html-extension.svg)
 
@@ -22,3 +24,11 @@ export default {
   plugins: [removeHtmlExtensionPlugin()],
 }
 ```
+
+## Contribution
+
+See [Contributing Guide](https://github.com/azat-io/azat-io/blob/main/contributing.md).
+
+## License
+
+MIT
