+++ tsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "types": ["node"],
+    "types": ["node", "@vuepress/client/types"],
     "forceConsistentCasingInFileNames": true,
     "importsNotUsedAsValues": "error",
     "moduleResolution": "nodenext",
