+++ vuepress.config.ts
@@ -1,8 +1,10 @@
 import { removeHtmlExtensionPlugin } from 'vuepress-plugin-remove-html-extension'
+import { umamiAnalyticsPlugin } from 'vuepress-plugin-umami-analytics'
 import postcssOklabFunction from '@csstools/postcss-oklab-function'
 import mdImageLazyLoading from 'markdown-it-image-lazy-loading'
 import { themeDataPlugin } from '@vuepress/plugin-theme-data'
 import { openGraphPlugin } from 'vuepress-plugin-open-graph'
+import { mermaidJsPlugin } from 'vuepress-plugin-mermaid-js'
 import { viteBundler } from '@vuepress/bundler-vite'
 import { shikiPlugin } from '@vuepress/plugin-shiki'
 import { defineUserConfig } from '@vuepress/cli'
@@ -15,9 +17,7 @@ import { loadTheme } from 'shiki'
 
 import { copyCodePlugin } from './plugins/copy-code/index.js'
 import { sitemapPlugin } from './plugins/sitemap/index.js'
-import { mermaidPlugin } from './plugins/mermaid/index.js'
 import { postsPlugin } from './plugins/posts/index.js'
-import { umamiPlugin } from './plugins/umami/index.js'
 import en from './locales/en.json'
 import ru from './locales/ru.json'
 
@@ -117,9 +117,9 @@ export default defineUserConfig({
     sitemapPlugin({
       hostname: 'https://azat.io',
     }),
-    mermaidPlugin(),
+    mermaidJsPlugin(),
     postsPlugin(),
-    umamiPlugin({
+    umamiAnalyticsPlugin({
       id: 'b198bd05-a70f-4a22-a46e-43908060c5a7',
       src: 'https://analytics.azat.io/umami.js',
     }),
